backend/scripts/requirements.txt
firebase-admin>=6.5.0
python-dotenv>=1.0.1
chromadb>=0.5.0
sentence-transformers>=2.2.2
torch>=2.0.0
accelerate>=0.21.0